Split Plugin logic from helpers
